Kernel/SysFS: Split bulky SysFSUSB file into two separate class files
